requirements.txt
Django>=4.2
djangorestframework
channels
channels-redis
